requirements.txt
wandb
mlflow
pandas
tqdm
huggingface_hub
boto3
asyncio
fastapi
google-cloud-storage